overall fixes

+++ application/views/back/pages/slider/index.php
@@ -28,11 +28,9 @@
                   <td class="align-middle"><?php echo $value->title; ?></td>
                   <td class="text-right">
                     <a href="<?php echo base_url(); ?>panel/<?php echo $this->uri->segment(2); ?>/form/update/<?php echo $value->id; ?>" class="btn btn-sm btn-info"><i class="icon ion-compose mg-r-10"></i> Edytuj</a>
-                    <!-- 
-                      <a href="<?php echo base_url(); ?>panel/settings/delete/<?php echo $this->uri->segment(2); ?>/<?php echo $value->id; ?>" class="btn btn-sm btn-secondary" 
-                      onclick="return confirm('Czy na pewno chcesz usunąć <?php echo $value->title; ?>? #<?php echo $value->id; ?>')" >
-                        <i class="fa fa-close mg-r-10"></i> Usuń
-                      </a> -->
+                    <a href="<?php echo base_url(); ?>panel/settings/delete/<?php echo $this->uri->segment(2); ?>/<?php echo $value->id; ?>" class="btn btn-sm btn-secondary" onclick="return confirm('Czy na pewno chcesz usunąć <?php echo $value->title; ?>? #<?php echo $value->id; ?>')">
+                      <i class="fa fa-close mg-r-10"></i> Usuń
+                    </a>
                   </td>
                 </tr>
               <?php endforeach; ?>
